chatbot/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "django-ipware==7.0.1",
    "django-redis==5.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "pydantic>=1.7.3",
    "python-dotenv==1.1.0",
    "langchain-core",
    "langchain-openai",
    "psycopg2-binary",
    "redis==5.2.1",
    "chromadb>=1",
    "langchain_community",
    "django-allauth",
    "cryptography",
    "dj_rest_auth"

]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.11.1"
]
